tests - Fix issue with coveralls gem
coveralls depends on term-ansicolor gem to colorize the output.
`Term` is defined as a Module in that gem. That conflicts with
the Term model we define within our project.

Fix for this is to not use term-ansicolor:

Coveralls::Output.no_color = true

The no_color option will be added to coveralls gem in v0.7.1, which
isn't out yet. Getting gem from the github repo until then..